Add json locale format
