+++ lib/tasks/locale.rake
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
-require "react_on_rails/locales_to_js"
+require "react_on_rails/locales/base"
+require "react_on_rails/locales/to_js"
+require "react_on_rails/locales/to_json"
 require "active_support"
 
 namespace :react_on_rails do
@@ -10,6 +12,10 @@ namespace :react_on_rails do
     the "ReactOnRails.configuration.i18n_dir".
   DESC
   task locale: :environment do
-    ReactOnRails::LocalesToJs.new
+    ReactOnRails::Locales::ToJs.new
+  end
+
+  task locale_json: :environment do
+    ReactOnRails::Locales::ToJson.new
   end
 end
